# Fix Issue 8156 - Optimize string comdats name generation (Windows)