~Replace md5 checksum with the already calculated hash in `stringTab`~, and don't call `mangleToBuffer` in advance if we know the string is long enough .

Test case:
```
void main()
{
   auto i = import("bigfile"); // tested with 102mb file
}
```

Tested using `-c -J. test.d --DRT-gcopt=profile:1 -lowmem` on QEMU.

Old behavior:

`-m32` switch:
Time: 3m07s,  RAM: 779mb    <-- really slow, mostly handled by #11554
`-m32mscoff` switch:
Time: 4.7s,  RAM: 779mb
`-m64` switch:
Time: 4.7s,  RAM: 779mb

New behavior (DMD built with #11554 patch too):

`-m32` switch:
Time: 2.1s,  RAM: 319mb
`-m32mscoff` switch:
Time: 2s,  RAM: 319mb
`-m64` switch:
Time: 2s,  RAM: 319mb