@ibuclaw 2020-08-12
> Old behavior:
> 
> `-m32` switch:
> Time: 3m07s,  RAM: 779mb    <-- really slow, mostly handled by #11554
> `-m32mscoff` switch:
> Time: 4.7s,  RAM: 779mb
> `-m64` switch:
> Time: 4.7s,  RAM: 779mb
> 
> New behavior (DMD built with #11554 patch too):
> 
> `-m32` switch:
> Time: 0.7s,  RAM: 319mb
> `-m32mscoff` switch:
> Time: 0.6s,  RAM: 319mb
> `-m64` switch:
> Time: 0.6s,  RAM: 319mb

I think your benchmark should have multiple modules that import bigfile. Compile separately and link together. There sh be no observed executable size regressions.