@WalterBright 2020-08-12
I have serious reservations about this. The hash in `stringTab` is optimized for speed, not uniqueness. Any collisions are handled by a linear search.

The purpose of the md5 hash is to ensure uniqueness - not just within the object file, but across all the object files linked together (so identical strings share storage).

The hash has to be unique. Using `stringTab`'s hash does not meat this requirement.

I recommend instead looking at speeding up the md5 calculation. Here is some work done on speeding up sha:

https://github.com/dlang/phobos/pull/7534