@BorisCarvajal 2020-08-13
@WalterBright That concern is valid, so I've discarded the md5 replacement, still the other changes(*) bring like 55% speed up (DMD not in release mode though).

(changes):
* Don't call mangleToBuffer in advance if we know the string is long enough.
* Checksum over the original string not the mangled one.